Company, pep, serializers, models improvements. Merge company and pep serializers for resolve circular imports
