--- business_register/serializers/pep_serializers.py
@@ -1,91 +1,0 @@
-from rest_framework import serializers
-from business_register.models.company_models import Founder
-from business_register.models.pep_models import Pep, PepRelatedPerson, CompanyLinkWithPep
-from business_register.serializers.company_serializers import (CountFoundedCompaniesSerializer,
-                                                               CompanyShortSerializer)
-
-
-class RelatedPersonSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = PepRelatedPerson
-        fields = (
-            'id', 'fullname', 'fullname_eng', 'relationship_type', 'relationship_type_eng',
-            'is_pep', 'start_date', 'confirmation_date', 'end_date'
-        )
-
-
-class CompanyLinkWithPepSerializer(serializers.ModelSerializer):
-    company = CompanyShortSerializer()
-
-    class Meta:
-        model = CompanyLinkWithPep
-        fields = ('company', 'company_name_eng', 'company_short_name_eng',
-                  'relationship_type', 'relationship_type_eng', 'start_date', 'end_date',
-                  'is_state_company')
-
-
-class CompanyDetailLinkWithPepSerializer(serializers.ModelSerializer):
-    company = CountFoundedCompaniesSerializer()
-
-    class Meta:
-        model = CompanyLinkWithPep
-        fields = (
-            'company', 'company_name_eng', 'company_short_name_eng', 'relationship_type',
-            'relationship_type_eng', 'start_date', 'end_date', 'is_state_company'
-        )
-
-
-class PepDetailSerializer(serializers.ModelSerializer):
-    related_persons = RelatedPersonSerializer(many=True)
-    related_companies = serializers.SerializerMethodField()
-    # other companies founded by persons with the same fullname as pep
-    check_companies = serializers.SerializerMethodField()
-
-    def get_related_companies(self, obj):
-        queryset = obj.related_companies.select_related('company', 'company__company_type', 'company__status').all()
-        return CompanyDetailLinkWithPepSerializer(queryset, many=True).data
-
-    def get_check_companies(self, pep):
-        pep_name = ' '.join([x for x in [pep.last_name, pep.first_name, pep.middle_name] if x])
-        founder_of = Founder.objects.filter(
-            name__contains=pep_name
-        ).select_related('company')
-        if not len(founder_of):
-            return []
-        check_companies = []
-        related_companies_id = pep.related_companies.values_list('company_id', flat=True)
-        for founder in founder_of:
-            if founder.company_id not in related_companies_id:
-                check_companies.append(
-                    CountFoundedCompaniesSerializer(founder.company).data
-                )
-        return check_companies
-
-    class Meta:
-        model = Pep
-        fields = ('id', 'fullname', 'fullname_eng', 'fullname_transcriptions_eng', 'last_job_title',
-                  'last_job_title_eng', 'last_employer', 'last_employer_eng', 'is_pep', 'pep_type',
-                  'pep_type_eng', 'url', 'info', 'info_eng', 'sanctions', 'sanctions_eng',
-                  'criminal_record', 'criminal_record_eng', 'assets_info', 'assets_info_eng',
-                  'criminal_proceedings', 'criminal_proceedings_eng', 'wanted', 'wanted_eng',
-                  'date_of_birth', 'place_of_birth', 'place_of_birth_eng', 'is_dead',
-                  'termination_date', 'reason_of_termination', 'reason_of_termination_eng',
-                  'related_persons', 'related_companies', 'check_companies',
-                  )
-
-
-class PepListSerializer(serializers.ModelSerializer):
-    related_persons = RelatedPersonSerializer(many=True)
-    related_companies = CompanyLinkWithPepSerializer(many=True)
-
-    class Meta:
-        model = Pep
-        fields = ('id', 'fullname', 'fullname_eng', 'fullname_transcriptions_eng', 'last_job_title',
-                  'last_job_title_eng', 'last_employer', 'last_employer_eng', 'is_pep', 'pep_type',
-                  'pep_type_eng', 'url', 'info', 'info_eng', 'sanctions', 'sanctions_eng',
-                  'criminal_record', 'criminal_record_eng', 'assets_info', 'assets_info_eng',
-                  'criminal_proceedings', 'criminal_proceedings_eng', 'wanted', 'wanted_eng',
-                  'date_of_birth', 'place_of_birth', 'place_of_birth_eng', 'is_dead',
-                  'termination_date', 'reason_of_termination', 'reason_of_termination_eng',
-                  'related_persons', 'related_companies',
-                  )
