+++ business_register/views/pep_views.py
@@ -3,7 +3,7 @@ from rest_framework import viewsets
 from rest_framework.permissions import AllowAny
 from business_register.filters import PepFilterSet
 from business_register.models.pep_models import Pep
-from business_register.serializers.pep_serializers import PepListSerializer, PepDetailSerializer
+from business_register.serializers.company_and_pep_serializers import PepListSerializer, PepDetailSerializer
 from data_ocean.views import CachedViewMixin
 from rest_framework.filters import SearchFilter
